Generate floats, not ints

+++ src/data.jl
@@ -54,7 +54,7 @@ function generatedata(
       x = x * chol.L
     end
   elseif x_type == :binary
-    x = Int.(sprand(Bool, n, p, Float64(x_density)))
+    x = Float64.(sprand(Bool, n, p, Float64(x_density)))
 
     if ρ != 0
       for j in 2:p
